mailbox: meson: Add Amlogic MHU mailbox driver
The mailbox driver is needed to communicate with the system controller,
which controls wol and power options from u-boot.

The driver has been ported from the vendor kernel for odroid[1] devices.
Also changes are made to the realtek phy driver to enable or disable wol
via the scpi protocol api, when using ethtool.

[1] https://github.com/hardkernel/linux

Signed-off-by: Lutz Koschorreck <theleks@ko-hh.de>